+++ tests/data/addon1/__openerp__.py
@@ -1,6 +1,9 @@
 {
     'name': 'addon 1',
-    'summary': 'addon 1 summary',
+    'summary': '''
+        addon 1 
+        summary
+    ''',
     'version': '8.0.1.0.0',
     'author': 'ACSONE SA/NV, '
               'Odoo Community Association (OCA)',