+++ tests/test_install_requires.py
@@ -4,7 +4,9 @@
 import os
 import unittest
 
-import setuptools_odoo
+from setuptools_odoo.core import (
+    get_install_requires_odoo_addon, get_install_requires_odoo_addons
+)
 
 from . import DATA_DIR
 
@@ -13,7 +15,7 @@ class TestInstallRequires(unittest.TestCase):
     """ Test the install_requires... public api """
 
     def test_addons_dir(self):
-        r = setuptools_odoo.get_install_requires_odoo_addons(DATA_DIR)
+        r = get_install_requires_odoo_addons(DATA_DIR)
         self.assertEqual(set(r), set([
             'astropy',
             # we have a mix of addons version, so two versions of Odoo
@@ -26,12 +28,12 @@ class TestInstallRequires(unittest.TestCase):
 
     def test_addon1(self):
         addon_dir = os.path.join(DATA_DIR, 'addon1')
-        r = setuptools_odoo.get_install_requires_odoo_addon(addon_dir)
+        r = get_install_requires_odoo_addon(addon_dir)
         self.assertEqual(r, ['odoo>=8.0a,<9.0a'])
 
     def test_addon2(self):
         addon_dir = os.path.join(DATA_DIR, 'addon2')
-        r = setuptools_odoo.get_install_requires_odoo_addon(addon_dir)
+        r = get_install_requires_odoo_addon(addon_dir)
         self.assertEqual(r, ['odoo8-addon-addon1',
                              'odoo>=8.0a,<9.0a',
                              'python-dateutil'])